fix(mcp): fix mcp gateway config (#1744)
