+++ spring-ai-alibaba-mcp/spring-ai-alibaba-mcp-gateway/spring-ai-alibaba-mcp-gateway-nacos/src/main/java/com/alibaba/cloud/ai/mcp/gateway/nacos/provider/NacosMcpAsyncGatewayToolsProvider.java
@@ -16,13 +16,14 @@
 
 package com.alibaba.cloud.ai.mcp.gateway.nacos.provider;
 
-import com.alibaba.cloud.ai.mcp.gateway.nacos.callback.DynamicNacosToolCallback;
-import com.alibaba.cloud.ai.mcp.gateway.nacos.definition.NacosMcpGatewayToolDefinition;
+import com.alibaba.cloud.ai.mcp.gateway.core.McpGatewayToolDefinition;
+import com.alibaba.cloud.ai.mcp.gateway.core.McpGatewayToolManager;
+import com.alibaba.cloud.ai.mcp.gateway.nacos.callback.NacosMcpGatewayToolCallback;
 import io.modelcontextprotocol.server.McpAsyncServer;
 import org.springframework.ai.mcp.McpToolUtils;
 import org.springframework.ai.tool.ToolCallback;
 
-public class NacosMcpAsyncGatewayToolsProvider implements NacosMcpGatewayToolsProvider {
+public class NacosMcpAsyncGatewayToolsProvider implements McpGatewayToolManager {
 
 	private final McpAsyncServer mcpAsyncServer;
 
@@ -31,14 +32,14 @@ public class NacosMcpAsyncGatewayToolsProvider implements NacosMcpGatewayToolsPr
 	}
 
 	@Override
-	public void addTool(final NacosMcpGatewayToolDefinition toolDefinition) {
+	public void addTool(final McpGatewayToolDefinition toolDefinition) {
 		try {
 			removeTool(toolDefinition.name());
 		}
 		catch (Exception e) {
 			// Ignore exception
 		}
-		ToolCallback toolCallback = new DynamicNacosToolCallback(toolDefinition);
+		ToolCallback toolCallback = new NacosMcpGatewayToolCallback(toolDefinition);
 		mcpAsyncServer.addTool(McpToolUtils.toAsyncToolSpecification(toolCallback)).block();
 	}
